#6 Placing new & saved entities after position confirmation; added clear map button; saving modified map after pressing Save button

+++ Assets/Scripts/Menu/EditEntityBehaviourListController.cs
@@ -24,7 +24,7 @@ public class EditEntityBehaviourListController: MonoBehaviour {
         }
 
         included = !included;
-        behaviourButtons[(int)behaviour].SetIncluded(included);
+        behaviourButtons[behaviour].SetIncluded(included);
         didChooseBehaviours?.Invoke(includedBehaviours);
     }
 }
